backend/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.30.0
flask-cors==4.0.0
flask-jwt-extended==4.5.3
python-dotenv==1.0.0
requests==2.31.0
psycopg2-binary==2.9.9
pytest==7.4.3
gunicorn==21.2.0
